+++ src/mkvtoolnix-gui/merge/source_file.h
@@ -55,6 +55,7 @@ public:
   virtual bool isAdditionalPart() const;
   virtual bool isPlaylist() const;
   virtual bool hasRegularTrack() const;
+  virtual bool hasVideoTrack() const;
 
   virtual void saveSettings(Util::ConfigFile &settings) const;
   virtual void loadSettings(MuxConfig::Loader &l);